+++ frontend/src/components/common/ClubItem.js
